requirements.txt
d2to1>=0.2.10,<0.3
Flask>=0.9
Flask-Gravatar
iso8601
launchpadlib
oslo.config
pbr>=0.5.16,<0.6
psutil
pylibmc
pymongo
sh
six
